Function spirit_log::init

pub fn init()

Initialize the global state.

This installs a global logger that can be replaced at runtime and sets a panic hook to also log panics (see log_panics).

This allows calling install later on.

It is needed only if the crate is used in the manual way. This is taken care of if used through the Pipeline.